requirements.txt
APScheduler==3.10.1
asgiref==3.7.1
Django==4.2.1
django-apscheduler==0.6.2
django-cors-headers==4.0.0
djangorestframework==3.14.0
geographiclib==2.0
geopy==2.3.0
psycopg2-binary==2.9.6
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tzlocal==5.0.1
